Add new RPC version block for job info
This is part of a series of development changes for adding the feature
of changing the mail-user and mail-type of jobs. Currently, these are
not given to clients as part of the job info, so the protocol for job
info needs to be changed. This commit is to follow the procedure for
doing that.

Bug 7780